+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNNWNVQ1X4FTTCJR06QBY80G"
+        "01KNP1EM6WQZ51RDT1FFP7Z88E"
     )
     r"""The unique identifier of the data source"""