+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNNWNVQ0YTHEKX3KN3WHZ69Y"
+        "01KNP1EM6VFEJZQ5DP9ZJ76CNS"
     )
     r"""The unique identifier of the data source"""