+++ packages/orq-rc/RELEASES.md
@@ -6909,3 +6909,13 @@ Based on:
 - [python v4.7.0-rc.51] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.51] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.51 - packages/orq-rc
+
+## 2026-04-08 07:58:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.52] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.52 - packages/orq-rc